Support Step source-path option.
Calling valohai.prepare() with multifile=True will enable dynamic
commands, which set the file to execute from the source-path Step option.
Note that valohai-utils will never change the existing command, so
enabling this on an existing Step will require manual editing of the
command.

source-path is not otherwise set--doing so is backwards incompatible, so
using it requires the entire stack to upgrade valohai-yaml.

This enables support for changing the file to invoke when updating
a step with several possible source files.