Add helper for handling execution output properties
The helper can be used as a context manager to set the metadata
properties for execution outputs. The user doesn't need to care about
the metadata file location or format; the helper takes care of that.

Metadata is automatically saved when exiting the context manager. A more
granular saving method can be implemented later if keeping all metadata
in memory proves an issue (tested with a million entries so should be ok
for pretty much all cases though).