Add shortcut for creating datasets with outputs
Instead of having to manually add the `valohai.dataset-versions`
property, add an optional parameter for adding the file to new dataset
versions. The helper merges the dataset version property to the other
properties.

When the file is created, print a summary to the execution logs, listing
the dataset versions created and how many output files were added to
each of them. This is helpful for checking the logs for correct output
handling.