Add incremental `add` method to properties helper
Move towards using `add` + separate method for adding files to a dataset
version, instead of doing everything in one method that overwrites
existing values.

This would be useful for users wanting to do things in bits instead of
preparing everything before setting properties.

The goal is to remove the `set` method altogether once the
transformation is complete.