Add separate method for adding files to datasets
This makes the user interface simpler: the user does not have to
construct the dataset versions URIs into a list (most use cases only
have one dataset anyway, so this is more convenient).

The interface is a bit more verbose (adding properties and every dataset
separately), but it is easier to understand, and works very well for the
most common use cases (just adding properties, or adding a file to one
dataset).