Preserve existing metadata from earlier context
Enables the user to process their metadata in several places (even
different scripts) within the execution.

Doing this slows things down a bit, as the metadata needs to be written
to a file and read back into an object again. But this shouldn't be an
issue in most cases, and if that is the bottleneck, only having one
metadata context is a quick solution.