Narrow IO type to one that supports writing bytes
Mypy complained that shutil.copyfileobj requires a SupportsWrite[AnyStr]
type, and even though destination was that, its type wasn't. Introduce a
type guard that assures the type checker that yes, we can write (bytes)
to the object.

This allows removing the `type: ignore` directives that Mypy really did
not like.