🏷️ Check types with Mypy (#151)
* Add mypy recipe for type checking

`make mypy` runs static type checks (requires `mypy` to be globally
installed).

If inside a virtual environment, point to the Python executable to
include the installed packages.

* Add Mypy to README

Instructions on running the type checking manually.

Mypy doesn’t really play nice with pre-commit, but that may still be
added to the recipes.